agents/addon/extension/search_python/requirements.txt
duckduckgo-search>=2.8.5
python-dotenv>=0.19.0
requests>=2.27.1